refactor: renamed new component to TokenLogo and moved to tokens folder
