+++ src/frontend/src/lib/components/tokens/TokenLogo.svelte
